+++ NEWS.md
@@ -1,3 +1,8 @@
+# Exametrika 0.11.6 on July 31, 2024.
+
+* bug fix for Item Total Correlation
+* New function called ItemStatistics
+
 # Exametrika 0.11.5 on Mar 30, 2024.
 
 * bug fix for plot labels for TRP/LRD